Added Catalog layer options
