+++ knowage-vue/src/modules/documentExecution/dashboard/widget/WidgetEditor/WidgetEditorTabs.vue
@@ -1,7 +1,7 @@
 <template>
     <TabView :active-index="activeIndex">
         <TabPanel v-if="propWidget && propWidget.type !== 'selection' && propWidget.type !== 'image'" :header="$t(propWidget.type === 'map' ? 'common.layers' : 'common.data')">
-            <MapWidgetLayersTab v-if="propWidget.type === 'map'" :prop-widget="propWidget" :datasets="datasets" :selected-datasets="selectedDatasets"></MapWidgetLayersTab>
+            <MapWidgetLayersTab v-if="propWidget.type === 'map'" :prop-widget="propWidget" :datasets="datasets" :selected-datasets="selectedDatasets" :layers="layers"></MapWidgetLayersTab>
             <WidgetEditorDataTab v-else :prop-widget="propWidget" :datasets="datasets" :selected-datasets="selectedDatasets" data-test="data-tab" @datasetSelected="$emit('datasetSelected', $event)"></WidgetEditorDataTab>
         </TabPanel>
         <TabPanel :header="$t('common.settings')">
@@ -13,6 +13,7 @@
                 :dashboard-id="dashboardId"
                 :html-gallery-prop="htmlGalleryProp"
                 :custom-chart-gallery-prop="customChartGalleryProp"
+                :layers="layers"
                 @settingChanged="$emit('settingChanged', $event)"
             ></WidgetEditorSettingsTab>
         </TabPanel>
@@ -25,11 +26,13 @@
  */
 import { defineComponent, PropType } from 'vue'
 import { IWidget, IDataset, IVariable, IGalleryItem } from '../../Dashboard'
+import { ILayer } from '@/modules/documentExecution/dashboard/interfaces/mapWidget/DashboardMapWidget'
 import TabView from 'primevue/tabview'
 import TabPanel from 'primevue/tabpanel'
 import WidgetEditorDataTab from './WidgetEditorDataTab/WidgetEditorDataTab.vue'
 import WidgetEditorSettingsTab from './WidgetEditorSettingsTab/WidgetEditorSettingsTab.vue'
 import MapWidgetLayersTab from './MapWidget/MapWidgetLayersTab.vue'
+import { AxiosResponse } from 'axios'
 
 export default defineComponent({
     name: 'widget-editor-tabs',
@@ -46,7 +49,19 @@ export default defineComponent({
     emits: ['datasetSelected', 'settingChanged'],
     data() {
         return {
-            activeIndex: 0
+            activeIndex: 0,
+            layers: [] as ILayer[]
+        }
+    },
+    created() {
+        if (this.propWidget.type === 'map') this.loadLayers()
+    },
+    methods: {
+        async loadLayers() {
+            await this.$http
+                .get(import.meta.env.VITE_RESTFUL_SERVICES_PATH + 'layers')
+                .then((response: AxiosResponse<any>) => (this.layers = response.data.root))
+                .catch(() => {})
         }
     }
 })
